Rewrite lineage calculation and run.sh
The lineage calculation is now non-recursive, and writes to a tsv file. For
this, a set of taxons for which the lineage was calculated is kept in memory.

run.sh was rewritten, so the invalid taxons are marked before doing any lineage
calculations. Now the lineages have to be calculated only once.